Safari-609.1.11.1/LayoutTests/fast/selectors/read-only-read-write-contenteditable-basics-expected.txt

Test the basic features of ":read-only", ":read-write" on elements with the contenteditable attribute. The definition is that ":read-write" is matches for "elements that are editing hosts or editable and are neither input elements nor textarea elements"

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.querySelectorAll("#test-block *")[0]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[0]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[1]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[1]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[2]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[3]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[4]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[4]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[5]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[5]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[6]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[7]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[8]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[8]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[9]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[9]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[10]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[11]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[12]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[12]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[13]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[13]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[14]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[15]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[16]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[16]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[17]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[17]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[18]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[19]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[20]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[20]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[21]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[21]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[22]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[22]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[23]).color is "rgb(0, 255, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[23]).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[24]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[24]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[25]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[25]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[26]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[26]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[27]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-block *")[27]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll("#test-block :read-write").length is 12
PASS document.querySelectorAll("#test-block :read-write")[0].id is "editing-host-1"
PASS document.querySelectorAll("#test-block :read-write")[1].id is "editable-block-1"
PASS document.querySelectorAll("#test-block :read-write")[2].id is "editing-host-2"
PASS document.querySelectorAll("#test-block :read-write")[3].id is "editable-block-2"
PASS document.querySelectorAll("#test-block :read-write")[4].id is "editing-host-3"
PASS document.querySelectorAll("#test-block :read-write")[5].id is "editable-block-3"
PASS document.querySelectorAll("#test-block :read-write")[6].id is "editing-host-4"
PASS document.querySelectorAll("#test-block :read-write")[7].id is "editable-block-4"
PASS document.querySelectorAll("#test-block :read-write")[8].id is "editing-host-5"
PASS document.querySelectorAll("#test-block :read-write")[9].id is "read-write-input"
PASS document.querySelectorAll("#test-block :read-write")[10].id is "editing-host-6"
PASS document.querySelectorAll("#test-block :read-write")[11].id is "read-write-textarea"
PASS document.querySelectorAll("#test-block :read-only").length is 16
PASS document.querySelectorAll("#test-block :read-only")[0].id is "non-editable-block-1"
PASS document.querySelectorAll("#test-block :read-only")[1].id is "non-editable-subblock-1"
PASS document.querySelectorAll("#test-block :read-only")[2].id is "non-editable-block-2"
PASS document.querySelectorAll("#test-block :read-only")[3].id is "non-editable-subblock-2"
PASS document.querySelectorAll("#test-block :read-only")[4].id is "non-editable-block-3"
PASS document.querySelectorAll("#test-block :read-only")[5].id is "non-editable-subblock-3"
PASS document.querySelectorAll("#test-block :read-only")[6].id is "non-editable-block-4"
PASS document.querySelectorAll("#test-block :read-only")[7].id is "non-editable-subblock-4"
PASS document.querySelectorAll("#test-block :read-only")[8].id is "read-only-input-1"
PASS document.querySelectorAll("#test-block :read-only")[9].id is "read-only-input-2"
PASS document.querySelectorAll("#test-block :read-only")[10].id is "read-only-input-3"
PASS document.querySelectorAll("#test-block :read-only")[11].id is "read-only-input-4"
PASS document.querySelectorAll("#test-block :read-only")[12].id is "read-only-textarea-1"
PASS document.querySelectorAll("#test-block :read-only")[13].id is "read-only-textarea-2"
PASS document.querySelectorAll("#test-block :read-only")[14].id is "read-only-textarea-3"
PASS document.querySelectorAll("#test-block :read-only")[15].id is "read-only-textarea-4"
PASS successfullyParsed is true

TEST COMPLETE
